Moves all postControllers to same directory
